split interrupts out from GPIO
